requirements.txt
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
bnlp-toolkit==3.3.2
captum==0.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
datasets==2.9.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
emoji==1.7.0
evaluate==0.4.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.0.3
gdown==4.7.1
gensim==4.3.1
huggingface-hub==0.12.1
idna==3.4
imbalanced-learn==0.11.0
imblearn==0.0
inquirerpy==0.3.4
ipykernel==6.21.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter-ydoc==0.2.2
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.6.0
jupyter_server_fileid==0.7.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.19.0
kaggle==1.5.13
kiwisolver==1.4.4
langdetect==1.0.9
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
multidict==6.0.4
multiprocess==0.70.14
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
nlpaug==1.1.11
nltk==3.8.1
normalizer @ git+https://github.com/csebuetnlp/normalizer@d80c3c484e1b80268f2b2dfaf7557fe65e34f321
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
overrides==7.3.1
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pfzy==0.3.4
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
prettytable==0.7.2
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.3
PySocks==1.7.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-json-logger==2.0.6
python-slugify==8.0.1
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sklearn-crfsuite==0.3.6
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
tabulate==0.9.0
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.1
transformers-interpret==0.10.0
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.14
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.7
xxhash==3.2.0
y-py==0.5.9
yarl==1.8.2
ypy-websocket==0.8.2
